* replaced gson with jackson * tpc communication with json * slave can send commands to the server * slave prints the song info

+++ core/src/main/java/at/rseiler/jango/core/player/MPlayer.java
@@ -22,7 +22,7 @@ public class MPlayer {
     /**
      * Start to play the songs from the station.
      */
-    public CompletableFuture<SongData> play(SongData songData, long songTime) {
+    public CompletableFuture<SongData> play(SongData songData, double songTime) {
         if (future != null && isAlive()) {
             future.cancel(true);
             stop();
